Make project compile: rewrite VMs, add MainWindow, fix DI
The AI-generated project had broken interfaces, missing files, missing
NuGet packages and ~25 missing AppConfig properties. Rebuilt the layer
between the working Core/Data services and the existing XAML so the
project now builds with 0 errors.

Highlights:
- Add Microsoft.Extensions.DependencyInjection package
- Delete broken ViewModels/Interfaces.cs (signatures didn't match impls)
- Add MainWindow.xaml/.cs with TabControl wiring all 5 tabs
- Add code-behind .xaml.cs for every tab (required by WPF)
- Add TranslationEngine: Capture -> OCR -> Translate -> Overlay pipeline
- Add HotkeyParser ("Ctrl+Shift+T" -> mod/vk) and TaskExtensions helpers
- Rewrite MainViewModel/SettingsViewModel/HistoryViewModel against the
  real concrete services (no interfaces); simpler tab XAMLs to match
- Extend AppConfig with hotkey strings and SourceLanguage
- Fix App.xaml.cs DI: register concrete types, supply appDataPath to
  CacheRepository and HistoryRepository, show MainWindow manually
- Fix OcrService: convert PNG bytes -> SoftwareBitmap, compute line
  bounds from word bounds (OcrLine has no BoundingRect)
- Fix HistoryRepository.StartSession (use SELECT last_insert_rowid())
- Fix TranslationService: drop unsupported TranslateTextOptions arg
- Fix WindowPickerService uint -> int cast for Process.GetProcessById
- Add .gitignore for bin/obj

+++ ErneyTranslateTool/Core/WindowPickerService.cs
@@ -51,7 +51,7 @@ public class WindowPickerService
             Process? process = null;
             try
             {
-                process = Process.GetProcessById(processId);
+                process = Process.GetProcessById((int)processId);
             }
             catch
             {
